Integrate with database and change ids type

+++ src/main/kotlin/com/gamerfinder/gamerfinder/dtos/input/CreateJoinRequestInput.kt
@@ -1,6 +1,6 @@
 package com.gamerfinder.gamerfinder.dtos.input
 
 data class CreateJoinRequestInput(
-    val playerId: Int,
+    val playerId: Long,
     val message: String? = null,
 )
